Change FTPClient module to include the client and object modules
